requirements.txt
boto3
botocore
python-ffmpeg
pydantic